% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_stand.R
\name{cpue_stand}
\alias{cpue_stand}
\title{CPUE standardization  for all subsets of possible variables}
\usage{
cpue_stand(
  variables,
  control_variables,
  data,
  dir.models,
  model_filename,
  error_dist = tw(link = "log"),
  stand_type = "gam"
)
}
\arguments{
\item{variables}{A vector of variables we want to consider}

\item{control_variables}{A formula in character form that specifies the response, as a function
of whatever control variables you want to include in all your models}

\item{data:}{A dataframe of the data you want to use}

\item{dir.models:}{The directory to your output folder}

\item{model_filename:}{The filename you want to give to your output object}

\item{error_dist:}{The error distribution you want to use for your standardization model. The default is to use a Tweedie distribution}

\item{stand_type:}{CPUE standardization model type. The default is to use mgcv to employ GAMs. Another option is to use GLMs}
}
\value{
A list object of models gets outputted into the environment and into the output directory
}
\description{
CPUE standardization  for all subsets of possible variables
}
\examples{
# Set working directory
dir.mod.output <- file.path(here(), "output")

# Create an object with possible variables we want to consider
possible_variables <- c("s(doy)", "te(AdjLon,lat2, bs = c('tp', 'tp'))",
"s(bottom_depth)",  "te(AdjLon, lat2, by = gear_descrip, bs = c('tp', 'tp'))", "area")

# Control variables we want to keep in every model
control_variables <- "weight ~ -1 + year + vessel_length_factor + offset(log(total_hooks_pots)) + type + gear_descrip +"

# Run function here!
cpue_stand(variables = possible_variables, control_variables = control_variables,
data = data, dir.models = dir.mod.output, model_filename = "cpue_standardization_allgear.RData",
error_dist = tw(link = "log"))
}
