% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_5fold_metrics.R
\name{extract_5fold_metrics}
\alias{extract_5fold_metrics}
\title{Extracts metrics from 5-fold cross validated models}
\usage{
extract_5fold_metrics(
  variables,
  model_object,
  data,
  dir.models,
  metric_filename
)
}
\arguments{
\item{variables}{A vector of variables we want to consider}

\item{modeL_boject}{The model object we want to extract metrics from}

\item{data:}{A dataframe of the data you want to use}

\item{dir.models:}{The directory to your output folder}

\item{metric_filename:}{The filename you want to give to your output object (metrics of CV; i.e., RMSE, MAE, R2)}
}
\value{
CSVs of out-of-sample metrics derived from the 5-fold cross validation process
}
\description{
Extracts metrics from 5-fold cross validated models
}
\examples{
# Load in 5 fold CV model object
load(file.path(dir.mod.output, "cpue_stand_5foldcv.RData"))

# Extract 5 fold CV metrics here
extract_metrics_5foldcv(variables = possible_variables,
                        model_object = cpue_stand_5foldcv,
                        data = data,
                        dir.models = dir.mod.output,
                        metric_filename =  "allgear_cv_metrics.csv")

}
