% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_kfold_metrics.R
\name{extract_kfold_metrics}
\alias{extract_kfold_metrics}
\title{Extracts metrics from 5-fold cross validated models}
\usage{
extract_kfold_metrics(
  variables,
  model_object,
  data,
  dir.models,
  metric_filename,
  folds,
  seed
)
}
\arguments{
\item{variables}{A vector of variables we want to consider}

\item{model_object}{The model object we want to extract metrics from}

\item{dir.models}{The directory to your output folder}

\item{metric_filename}{The filename you want to give to your output object (metrics of CV; i.e., RMSE, MAE, R2)}

\item{folds}{Number of fold we want to extract metrics out of}

\item{seed}{Seed used for k_fold_cv function for random splits}

\item{data:}{A dataframe of the data you want to use}
}
\value{
CSVs of out-of-sample metrics derived from the 5-fold cross validation process
}
\description{
Extracts metrics from 5-fold cross validated models
}
\examples{
# Load in k fold CV model object
load(file.path(dir.mod.output, "cpue_stand_5foldcv.RData"))

# Extract k fold CV metrics here
extract_metrics_5foldcv(variables = possible_variables,
                        model_object = cpue_stand_5foldcv,
                        data = data,
                        dir.models = dir.mod.output,
                        metric_filename =  "allgear_cv_metrics.csv",
                        folds = 5, seed = 666)

}
