% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/five_foldCV.R
\name{five_foldCV}
\alias{five_foldCV}
\title{Conducts 5-fold cross validation for cpue_standardization models}
\usage{
five_foldCV(
  data,
  variables,
  control_variables,
  dir.models,
  model_filename,
  error_dist = tw(link = "log"),
  stand_type = "gam",
  seed = 666
)
}
\arguments{
\item{variables}{A vector of variables we want to consider}

\item{control_variables}{A formula in character form that species the response, as a function
of whatever control variables you want to include in all your models}

\item{data:}{A dataframe of the data you want to use}

\item{dir.models:}{The directory to your output folder}

\item{model_filename:}{The filename you want to give to your output object}

\item{error_dist:}{The error distribution you want to use for your standardization model. The default is a Tweedie Distribution.}

\item{stand_type:}{CPUE standardization model type. The default is to use mgcv to employ GAMs. Another option is to use GLMs}

\item{seed:}{Set.seed for 5-fold cross validation}
}
\value{
A list object of 5-fold cross validated standardized CPUE models
}
\description{
Conducts 5-fold cross validation for cpue_standardization models
}
\examples{
# Set working directory
dir.mod.output <- file.path(here(), "output")

# Create an object with possible variables we want to consider
possible_variables <- c("s(doy)", "te(AdjLon,lat2, bs = c('tp', 'tp'))",
"s(bottom_depth)",  "te(AdjLon, lat2, by = gear_descrip, bs = c('tp', 'tp'))", "area")

# Control variables we want to keep in every model
control_variables <- "weight ~ -1 + year + vessel_length_factor + offset(log(total_hooks_pots)) + type + gear_descrip +"

# Run function here!
five_foldCV(variables = possible_variables, control_variables = control_variables,
data = data, dir.models = dir.mod.output, model_filename = "cpue_standardization_5foldCV.RData",
error_dist = tw(link = "log"))
}
